MVP reached

+++ controllers/api/commentRoutes.js
@@ -12,35 +12,39 @@ router.get("/", (req, res) => {
 });
 // add a new comment
 router.post("/", (req, res) => {
-  Comment.create({
-    comment_text: req.body.comment_text,
-    user_id: req.body.user_id,
-    post_id: req.body.post_id,
-  })
-    .then((dbCommentData) => res.json(dbCommentData))
-    .catch((err) => {
-      console.log(err);
-      res.status(400).json(err);
-    });
+  if (req.session) {
+    Comment.create({
+      comment_text: req.body.comment_text,
+      user_id: req.session.user_id,
+      post_id: req.body.post_id,
+    })
+      .then((dbCommentData) => res.json(dbCommentData))
+      .catch((err) => {
+        console.log(err);
+        res.status(400).json(err);
+      });
+  }
 });
 // delete a comment
-router.delete('/:id', (req, res) =>{
-  Comment.destroy({
-    where: {
-      id: req.params.id,
-    },
-  })
-    .then((dbCommentData) => {
-      if (!dbCommentData) {
-        res.status(404).json({ message: "No comment found with this id" });
-        return;
-      }
-      res.json(dbCommentData);
+router.delete("/:id", (req, res) => {
+  if (req.session) {
+    Comment.destroy({
+      where: {
+        id: req.params.id,
+      },
     })
-    .catch((err) => {
-      console.log(err);
-      res.status(500).json(err);
-    });
+      .then((dbCommentData) => {
+        if (!dbCommentData) {
+          res.status(404).json({ message: "No comment found with this id" });
+          return;
+        }
+        res.json(dbCommentData);
+      })
+      .catch((err) => {
+        console.log(err);
+        res.status(500).json(err);
+      });
+  }
 });
 
 module.exports = router;
